requirements.txt
Flask
requests
mysql-connector-python
python-dotenv
hashlib